requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
apturl==0.5.2
astroid==2.14.2
async-generator==1.10
attrs==22.2.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2022.12.7
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
dill==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.1.0
fasteners==0.14.1
future==0.18.2
h11==0.14.0
httplib2==0.14.0
idna==2.8
iniconfig==2.0.0
isort==5.12.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.9.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.7.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
outcome==1.2.0
packaging==23.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==3.0.0
pluggy==1.0.0
protobuf==3.6.1
py==1.11.0
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.16.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
PySocks==1.7.1
pytest==7.2.1
pytest-html==3.2.0
pytest-metadata==2.0.4
python-apt==2.0.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.21.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==4.8.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
systemd-python==234
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.14
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.3
webdriver-manager==3.8.5
wrapt==1.14.1
wsproto==1.2.0
xkit==0.0.0
